Refactor: Extract new fn `run_commands_in_parallel`
