+++ src/parallel/mod.rs
@@ -1,3 +1,6 @@
-pub(crate) mod async_executor;
-pub(crate) mod job_token;
+mod async_executor;
+mod command_runner;
+mod job_token;
 pub(crate) mod stderr;
+
+pub(crate) use command_runner::run_commands_in_parallel;
